requirements.txt
amqp==5.1.1
anyascii==0.3.2
arrow==1.3.0
async-timeout==4.0.3
Babel==2.13.0
beautifulsoup4==4.7.1
bibtexparser==1.1.0
billiard==3.6.4.0
binaryornot==0.4.4
blinker==1.6.2
celery==5.1.2
celery-once==3.0.1
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.3.0
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cookiecutter==1.7.3
Cython==3.0.2
dnspython==2.2.1
email-validator==1.1.3
et-xmlfile==1.1.0
eventlet==0.33.0
filedepot==0.10.0
Flask==1.1.4
Flask-BabelEx==0.9.4
Flask-Caching==1.10.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Script==2.0.6
Flask-Security==3.0.0
Flask-WTF==0.14.3
future==0.18.3
greenlet==3.0.0
html5lib==1.1
idna==3.4
ijson==2.4
isodate==0.6.1
itsdangerous==1.1.0
Jinja2==2.11.3
jinja2-time==0.2.0
joblib==1.3.2
kombu==5.3.2
lxml==4.9.3
Markdown==3.4.4
MarkupSafe==2.0.1
nltk==3.6.5
nose==1.3.7
numpy==1.26.0
openpyxl==3.1.2
oxrdflib==0.3.1
pandas==2.1.1
passlib==1.7.4
pip==22.3.1
poyo==0.5.0
prompt-toolkit==3.0.39
psutil==5.9.5
puremagic==1.14
PyJWT==2.8.0
pyoxigraph==0.3.19
pyparsing==3.1.1
pyshp==2.3.1
python-dateutil==2.8.2
python-slugify==8.0.1
pytidylib==0.3.2
pytz==2023.3.post1
rdflib==6.0.2
rdflib-jsonld==0.6.2
redis==5.0.1
redislite==6.2.899109
regex==2023.10.3
requests==2.31.0
requests-testadapter==0.3.0
sadi==0.5.7
scipy==1.11.3
sdd2rdf==1.3.1
setlr==0.2.19
setuptools==65.5.0
six==1.16.0
soupsieve==2.5
speaklater==1.3
text-unidecode==1.3
toposort==1.10
tqdm==4.66.1
types-python-dateutil==2.8.19.14
tzdata==2023.3
urllib3==2.0.6
vine==5.0.0
wcwidth==0.2.8
webencodings==0.5.1
WebOb==1.8.7
Werkzeug==1.0.1
whyis==2.1.2
WTForms==3.0.1
xlrd==2.0.1

mod-wsgi==4.9.0
spacy-transformers>=1.1.6,<1.2.0
en_core_web_lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.4.0/en_core_web_lg-3.4.0-py3-none-any.whl
spacy-experimental>=0.6.0,<0.7.0
spacy>=3.3.0,<4.0.0
